feat(consensus): compute root hash for validator_set
