+++ src/noir/consensus/validator.h
@@ -67,12 +67,7 @@ struct validator_set {
     return validators.size();
   }
 
-  bytes get_hash() {
-    for (auto val : validators) {
-      // todo
-    }
-    return bytes{};
-  }
+  bytes get_hash();
 
   bool has_address(const bytes& address) {
     for (const auto& val : validators) {
